step-2 회원가입 로직 구현
